Moved components to own directory
- snapshot tests for table and calculation form
